==> PIPELINE/4_CROSS_VALIDATION.Rmd <==

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


# 1. CROSS VALIDATION


```{r echo=FALSE}

## LOAD THE DATA

WD = '/Users/Remi/Dropbox/1_WORK/2015/4_PROJECTS/1_SOFTWARE/IMAGEJ_LIB/CELL SIGNALING ANALYSIS/CODE/2_CLASSIFY/BORDERCELLS/PIPELINE/'

setwd(WD)


# load the cross validation data-set
d <- read.delim("4_DATA/20170301_EARLY_CROSSVALID.txt")

# select only a specific experimental condition 
d = d[d$FILES_KEY=='CT_24',]

# load theta , yes bias value 
d_theta <- read.delim("4_DATA/OUT_features_theta.txt", header=F)
d_theta = data.matrix(d_theta)


# load mean, no bias value 
d_resc_mn <- read.delim("3_DATA/3_data_training_param_MEAN.txt")
d_resc_mn = data.matrix(d_resc_mn)
l = dim(d_resc_mn)[1]
d_resc_mn = d_resc_mn[1:l-1] # remove the last value which was ISEDGE=1 o 0 (value show the mean) 

# load stdev, no bias value 
d_resc_sd <- read.delim("3_DATA/3_data_training_param_STDEV.txt")
d_resc_sd = data.matrix(d_resc_sd)
l = dim(d_resc_sd)[1]
d_resc_sd = d_resc_sd[1:l-1] # remove the last value which was ISEDGE=1 o 0 (value show the mean) 

```



## NORMALIZE VALUES
```{r}
# 'VORO.protein3.ROUND' 
# 'VORO.protein3.AR' 
# 'VORO.Feret.norm' = VORO.protein3.Feret / NUCL.dna.Feret
# 'VORO.Area.norm'  = VORO.protein3.Area / NUCL.dna.Area
# 'CYTO.Area.norm' = CYTO.protein3.Area / NUCL.dna.Area
# 'NUVO.distance.norm' = sqrt((x1-x2)^2+(y1+y2)^2) 

X = d[,1:length(d_theta)]

names(X)[1] = 'BIAS'
names(X)[2] = 'VORO.protein3.ROUND' 
names(X)[3] = 'VORO.protein3.AR' 
names(X)[4] = 'VORO.Feret.norm'
names(X)[5] = 'VORO.Area.norm'
names(X)[6] = 'CYTO.Area.norm'
names(X)[7] = 'NUVO.distance.norm'

X[,1] = 1
X[,2] = d$VORO.SMAD.RND
X[,3] = d$VORO.SMAD.AR
X[,4] = d$VORO.SMAD.F/d$NUCL.NUCL.F
X[,5] = d$VORO.SMAD.A/d$NUCL.NUCL.A
X[,6] = d$CY_ROI.SMAD.A/d$NUCL.NUCL.A
X[,7] = sqrt((d$NUCL.NUCL.X-d$VORO.SMAD.X)^2+(d$NUCL.NUCL.Y-d$VORO.SMAD.Y)^2)/d$NUCL.NUCL.F


```

## cross validation data set and classifier info and parameters

```{r echo=FALSE}

# feature names
cat('cross-validation data set: feature names\n ', names(X[,2:7]))

#number of raws
cat('cross-validation data set: number of raws\n', dim(X)[1])

#number of features
cat('cross-validation data set: number of features\n', dim(X)[2])

# normalization transformation
cat('normalization data transformation: log2(x+1) \n')

# rescaling mean values 
cat('rescaling mean values\n', d_resc_mn)

# rescaling stedv values 
cat('rescaling stdev values\n', d_resc_sd)

# Parameters logistic regression
cat('Parameters logistic regression (includes bias)\n', d_theta)


```


## Features distribution

```{r echo=FALSE}
par(mfrow=c(4,2))
par(mar = rep(2, 4))
dd_mtx =log(data.matrix(X, rownames.force = NA)+1,2)
n_col = dim(dd_mtx)[2]
for(i in 1:(n_col-1)){
  hist(dd_mtx[,i],100,main=names(X)[i], cex.axis=1, font.main=1, cex.main=1.3)
}

```


## 2. Classification probability 
```{r}

X1 = data.matrix(X)
n = dim(X1)[1]
m = dim(X1)[2]

# log transf
for(i in 1:n){
  X1[i,2:m] = log10(X1[i,2:m]+1) 
}

# rescaling
for(i in 1:n){
  X1[i,2:m] = (X1[i,2:m] - data.matrix(d_resc_mn))/data.matrix(d_resc_sd)
}

# calculate the classification 
H = X1 %*% data.matrix(d_theta)

# claculate the probability (sigma)
prob = 1/(1+exp(-H));

prob2 = round(prob,4)



```

## probability for cells to be at the edge vs interior
```{r echo=FALSE}
par(mfrow=c(1,2))
plot(prob2[prob2>0.5], ylim=c(0,1))
plot(prob2[prob2<0.5], ylim=c(0,1))

cat (' Number at the edge=', length(prob2[prob2>0.5]))
cat (' Number at the interior=', length(prob2[prob2<0.5]))
```


## Add the prob.edge column to the data set and save 
```{r}
d2 = d
n_col = dim(d2)[2]
d2[,n_col+1 ] = prob2
names(d2)[n_col+1] = "prob.edge"

write.table(d2, file = "5_DATA/5_data_classified_CT.txt", sep = "\t", quote=FALSE, row.names=FALSE, col.names = colnames(d2))

write.table(prob2, file = "5_DATA/5_classifier_prob_CT.txt", sep = "\t", quote=FALSE, row.names=FALSE, col.names = colnames(prob2))

```
